requirements.txt
bottle==0.13.2
Pillow==11.2.1
qrcode==8.1
Unidecode==1.3.8
zbarlight==3.0
